undo vmservicescrape reconcile function renaming
