+++ internal/controller/operator/factory/vlsingle/vlogs.go
@@ -293,7 +293,7 @@ func createOrUpdateVLogsService(ctx context.Context, rclient client.Client, cr, 
 	if !ptr.Deref(cr.Spec.DisableSelfServiceScrape, false) {
 		svs := buildVLogsScrape(cr, svc)
 		prevSvs := buildVLogsScrape(prevCR, prevSvc)
-		if err := reconcile.VMServiceScrapeForCRD(ctx, rclient, svs, prevSvs, &owner); err != nil {
+		if err := reconcile.VMServiceScrape(ctx, rclient, svs, prevSvs, &owner); err != nil {
 			return fmt.Errorf("cannot create serviceScrape for vlsingle: %w", err)
 		}
 	}
