+++ internal/controller/operator/factory/vmalert/vmalert.go
@@ -71,7 +71,7 @@ func createOrUpdateService(ctx context.Context, rclient client.Client, cr, prevC
 	if !ptr.Deref(cr.Spec.DisableSelfServiceScrape, false) {
 		svs := buildScrape(cr, svc)
 		prevSvs := buildScrape(prevCR, prevSvc)
-		if err := reconcile.VMServiceScrapeForCRD(ctx, rclient, svs, prevSvs, &owner); err != nil {
+		if err := reconcile.VMServiceScrape(ctx, rclient, svs, prevSvs, &owner); err != nil {
 			return fmt.Errorf("cannot create vmservicescrape: %w", err)
 		}
 	}
