+++ internal/controller/operator/factory/vmalertmanager/statefulset.go
@@ -146,7 +146,7 @@ func createOrUpdateAlertManagerService(ctx context.Context, rclient client.Clien
 	if !ptr.Deref(cr.Spec.DisableSelfServiceScrape, false) {
 		svs := buildScrape(cr, svc)
 		prevSvs := buildScrape(prevCR, prevSvc)
-		if err := reconcile.VMServiceScrapeForCRD(ctx, rclient, svs, prevSvs, &owner); err != nil {
+		if err := reconcile.VMServiceScrape(ctx, rclient, svs, prevSvs, &owner); err != nil {
 			return err
 		}
 	}
