+++ internal/controller/operator/factory/vtcluster/insert.go
@@ -335,7 +335,7 @@ func createOrUpdateVTInsertService(ctx context.Context, rclient client.Client, c
 	if !ptr.Deref(cr.Spec.Insert.DisableSelfServiceScrape, false) {
 		svs := buildVTInsertScrape(cr, svc)
 		prevSvs := buildVTInsertScrape(prevCR, prevSvc)
-		if err := reconcile.VMServiceScrapeForCRD(ctx, rclient, svs, prevSvs, &owner); err != nil {
+		if err := reconcile.VMServiceScrape(ctx, rclient, svs, prevSvs, &owner); err != nil {
 			return fmt.Errorf("cannot create VMServiceScrape for VTInsert: %w", err)
 		}
 	}
